code clean and standardization
following an enquiry regarding difficulty and it's effects, something struck
me about how we were handling the data in various places.
sometimes we used a static_cast<int> and sometimes we cast (int) and
sometimes we did neither.
i figured some kind of standard was in order, so i settled on the middle ground
and cast them all as (int)

+++ src/Battlescape/BattlescapeGame.cpp
@@ -1541,7 +1541,7 @@ BattleUnit *BattlescapeGame::convertUnit(BattleUnit *unit, std::string newType)
 
 	BattleUnit *newUnit = new BattleUnit(getRuleset()->getUnit(newType), FACTION_HOSTILE, _save->getUnits()->back()->getId() + 1, getRuleset()->getArmor(newArmor.str()));
 	
-	int difficulty = _parentState->getGame()->getSavedGame()->getDifficulty();
+	int difficulty = (int)(_parentState->getGame()->getSavedGame()->getDifficulty());
 	int divider = 1;
 	if (!difficulty)
 		divider = 2;
